exports refactor and tidying up structure
